feat: combine migrations and add image resource entity

+++ src/modules/users/users.service.ts
@@ -13,6 +13,8 @@ import { UserDto } from './dto/user.dto';
 import { ChangePasswordDto } from '../auth/dto/change-password.dto';
 import { compareSync } from 'bcrypt';
 import { GoogleUserDto } from '../auth/dto/google-user.dto';
+import { ImageResource } from '../image-resources/entities';
+import { createImageResourceDtoFromCloudinaryResAdapter } from '../image-resources/adapters';
 
 @Injectable()
 export class UsersService {
@@ -21,11 +23,14 @@ export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly usersRepository: Repository<User>,
+
+    @InjectRepository(ImageResource)
+    private readonly imageResourcesRepository: Repository<ImageResource>,
+
     private readonly cloudinaryService: CloudinaryService,
   ) {}
 
   async create(createUserDto: CreateUserDto) {
-    delete createUserDto.passwordConfirmation;
     createUserDto.password = hashPassword(createUserDto.password);
 
     const user = this.usersRepository.create(createUserDto);
@@ -36,64 +41,60 @@ export class UsersService {
   }
 
   async createFromGoogle(googleUser: GoogleUserDto) {
-    const user = await this.usersRepository.findOne({ where: { email: googleUser.email } });
-    if (user) {
-      if (!user.emailVerifiedAt) user.emailVerifiedAt = new Date();
+    const { email, username, picture } = googleUser;
 
-      if (!user.profilePhoto && googleUser.picture) {
-        try {
-          user.profilePhoto = await this.cloudinaryService.uploadImageFromUrl(
-            googleUser.picture,
-            user.username,
-            CloudinaryPresets.PROFILE_PHOTO,
-          );
-        } catch (error) {
-          console.log(error);
-          this.logger.error('No se pudo cargar la imagen desde la url');
-        }
-      }
+    // * Create or update the user entity
+    let user = await this.usersRepository.findOne({ where: { email } });
+    if (!user) user = this.usersRepository.create({ email: email, username: username });
 
-      return this.usersRepository.save(user);
+    // * If the user is not verified, set the emailVerifiedAt field
+    if (!user.emailVerifiedAt) user.emailVerifiedAt = new Date();
+    await this.usersRepository.save(user);
+
+    console.log('User:', user);
+
+    if (user.profilePhoto || !picture) return user;
+
+    // * Add google profile photo to the user
+    try {
+      const cloudResponse = await this.cloudinaryService.uploadImageFromUrl(
+        picture,
+        user.username,
+        CloudinaryPresets.PROFILE_PHOTO,
+      );
+
+      const imageDto = createImageResourceDtoFromCloudinaryResAdapter(cloudResponse);
+      if (!imageDto) throw new Error('Error creating image resource DTO');
+
+      const image = this.imageResourcesRepository.create(imageDto);
+      await this.imageResourcesRepository.save(image);
+
+      user.profilePhoto = image;
+      await this.usersRepository.save(user);
+    } catch (error) {
+      console.log(error);
+      this.logger.error('No se pudo cargar la imagen de usuario desde la url de gooogle');
     }
 
-    const newImage: CloudinaryImage | undefined = googleUser.picture
-      ? await this.cloudinaryService.uploadImageFromUrl(
-          googleUser.picture,
-          googleUser.username,
-          CloudinaryPresets.PROFILE_PHOTO,
-        )
-      : undefined;
-
-    const newUser = this.usersRepository.create({
-      email: googleUser.email,
-      username: googleUser.username,
-      profilePhoto: newImage,
-      emailVerifiedAt: new Date(),
-    });
-
-    return this.usersRepository.save(newUser);
+    return user;
   }
 
   findAll() {
     return this.usersRepository.find();
   }
 
-  async findOne(id: string): Promise<User> {
-    const user = await this.usersRepository.findOne({ where: { id }, relations: ['role'] });
-    return user;
+  async findOne(id: string): Promise<User | null> {
+    return this.usersRepository.findOne({ where: { id }, relations: ['role'] });
   }
 
-  async findOneWithSessionAndRole(id: string, sessionId: string): Promise<User> {
-    return this.usersRepository
-      .createQueryBuilder('user')
-      .innerJoinAndSelect('user.sessions', 'session')
-      .leftJoinAndSelect('user.role', 'role')
-      .where('user.id = :id', { id })
-      .andWhere('session.id = :sessionId', { sessionId })
-      .getOne();
+  async findOneWithSessionAndRole(id: string, sessionId: string): Promise<User | null> {
+    return this.usersRepository.findOne({
+      where: { id, sessions: { id: sessionId } },
+      relations: { role: true, sessions: true },
+    });
   }
 
-  async getFullUser(email: string): Promise<User> {
+  async getFullUser(email: string): Promise<User | null> {
     return this.usersRepository
       .createQueryBuilder('user')
       .leftJoinAndSelect('user.role', 'role')
@@ -106,7 +107,7 @@ export class UsersService {
     const user = await this.findOne(id);
     if (!user) throw new NotFoundException('User not found');
 
-    const { profilePhoto: currentProfilePath } = user;
+    const { profilePhoto: currentImage } = user;
     let image: CloudinaryImage | undefined;
 
     try {
@@ -118,16 +119,18 @@ export class UsersService {
       });
 
       // * If the image is not uploaded, throw an error
-      if (!image) throw new BadRequestException('Error uploading image');
 
       // * Update the user's profile photo path
-      user.profilePhoto = image;
+      const imageDto = createImageResourceDtoFromCloudinaryResAdapter(image);
+      if (!imageDto) throw new BadRequestException('Error uploading image');
+
+      user.profilePhoto = await this.imageResourcesRepository.save(imageDto);
       await this.usersRepository.save(user);
 
       // * Delete the old profile photo if it exists
 
-      if (currentProfilePath) {
-        this.cloudinaryService.destroyFile(currentProfilePath.publicId);
+      if (currentImage && currentImage.publicId) {
+        this.cloudinaryService.destroyFile(currentImage.publicId);
       }
 
       return new UserDto(user);
@@ -147,11 +150,10 @@ export class UsersService {
     if (!user) throw new NotFoundException('User not found');
 
     const { profilePhoto } = user;
-    if (!profilePhoto) throw new BadRequestException('User does not have a profile photo');
+    if (!profilePhoto || !profilePhoto.publicId) throw new BadRequestException('User does not have a profile photo');
 
     await this.cloudinaryService.destroyFile(profilePhoto.publicId);
-    user.profilePhoto = null;
-    await this.usersRepository.save(user);
+    await this.imageResourcesRepository.delete({ id: profilePhoto.id });
 
     return new UserDto(user);
   }
@@ -160,7 +162,7 @@ export class UsersService {
     const { password, newPassword } = changePasswordDto;
 
     const user = await this.getFullUser(email);
-    if (!user) throw new NotFoundException('User not found');
+    if (!user || !user.password) throw new NotFoundException('User not found');
     if (!compareSync(password, user.password)) throw new UnauthorizedException('Invalid password');
 
     user.password = hashPassword(newPassword);
